refactor: move stuff from IWsEndpoint.ts to AbstractWsEndpoint.ts
