+++ packages/network/src/connection/AbstractWsEndpoint.ts
@@ -1,3 +1,7 @@
+export class AbstractWsEndpoint {
+
+}
+
 export enum Event {
     PEER_CONNECTED = 'streamr:peer:connect',
     PEER_DISCONNECTED = 'streamr:peer:disconnect',
@@ -8,18 +12,13 @@ export enum Event {
 
 export enum DisconnectionCode {
     GRACEFUL_SHUTDOWN = 1000,
-    DUPLICATE_SOCKET = 1002,
-    NO_SHARED_STREAMS = 1000,
     MISSING_REQUIRED_PARAMETER = 1002,
-    DEAD_CONNECTION = 1002
 }
 
 export enum DisconnectionReason {
     GRACEFUL_SHUTDOWN = 'streamr:node:graceful-shutdown',
     DUPLICATE_SOCKET = 'streamr:endpoint:duplicate-connection',
     NO_SHARED_STREAMS = 'streamr:node:no-shared-streams',
-    MISSING_REQUIRED_PARAMETER = 'streamr:node:missing-required-parameter',
-    DEAD_CONNECTION = 'streamr:endpoint:dead-connection'
 }
 
 export class UnknownPeerError extends Error {
