+++ packages/network/src/connection/ClientWsEndpoint.ts
@@ -1,11 +1,11 @@
 import { EventEmitter } from 'events'
-import { DisconnectionCode, DisconnectionReason, Event, UnknownPeerError } from './IWsEndpoint'
 import WebSocket from 'ws'
 import { PeerInfo } from './PeerInfo'
 import { Metrics, MetricsContext } from '../helpers/MetricsContext'
 import { Logger } from '../helpers/Logger'
 import { Rtts } from '../identifiers'
 import { PingPongWs } from "./PingPongWs"
+import { DisconnectionCode, DisconnectionReason, Event, UnknownPeerError } from "./AbstractWsEndpoint"
 
 const staticLogger = new Logger(module)
