+++ packages/network/src/protocol/TrackerServer.ts
@@ -3,12 +3,12 @@ import { v4 as uuidv4 } from 'uuid'
 import { TrackerLayer, TrackerMessageType } from 'streamr-client-protocol'
 import { Logger } from '../helpers/Logger'
 import { decode } from '../helpers/MessageEncoder'
-import { Event as WsEndpointEvent } from '../connection/IWsEndpoint'
 import { StreamIdAndPartition } from '../identifiers'
 import { PeerInfo } from '../connection/PeerInfo'
 import { RtcSubTypes } from '../logic/RtcMessage'
 import { NameDirectory } from '../NameDirectory'
 import { ServerWsEndpoint } from "../connection/ServerWsEndpoint"
+import { Event as WsEndpointEvent } from "../connection/AbstractWsEndpoint"
 
 export enum Event {
     NODE_CONNECTED = 'streamr:tracker:send-peers',
